+++ src/components/ContractInfo.test.jsx
@@ -4,7 +4,7 @@ import ContractInfo from './ContractInfo';
 
 it('renders correctly', () => {
   const tree = renderer
-    .create(<ContractInfo accountAddress="0x0123" contractAddress="0x3210" network={3} />)
+    .create(<ContractInfo accountAddress="0x0123" accountBalance={1.123} contractAddress="0x3210" contractBalance={123} network={3} />)
     .toJSON();
   expect(tree).toMatchSnapshot();
 });
@@ -12,7 +12,7 @@ it('renders correctly', () => {
 
 it('no account address', () => {
   const tree = renderer
-    .create(<ContractInfo contractAddress="0x3210" network={3} />)
+    .create(<ContractInfo contractAddress="0x3210" contractBalance={123} accountBalance={0} network={3} />)
     .toJSON();
   expect(tree).toMatchSnapshot();
 });